improve templates; move toc into AppContent
