+++ docs/content/2.usage/4.template.md
@@ -28,13 +28,13 @@ A wrapper component that contains everything about the page, including page cont
 
 Display table of contents of the document. This component does not show on [fullscreen document](/usage/content#front-matter). This component shows inside the page on the right side of document content.
 
-### `<PageBottom>`
+### `<EditOnGithub>`
 
 Display information about the document, including the last modified date and a link to edit the document. This component shows inside the page and under the document content.
 
 ### `<PagePrevNext>`
 
-Display links to the next and previous pages. This component shows inside the `<AppPage>` and under the `<PageBottom>`.
+Display links to the next and previous pages. This component shows inside the `<AppPage>` and under the `<EditOnGithub>`.
 
 ### `<AlgoliaSearchBox>`
