+++ src/defaultTheme/components/organisms/AppPage.vue
@@ -1,7 +1,11 @@
 <template>
   <div class="flex flex-col w-full pt-0 xl:flex-row xl:pt-10">
-    <article class="flex-auto order-last min-w-0 px-4 sm:px-6 mt-4 xl:order-first sm:px-6 xl:mt-0">
+    <article class="flex-auto order-last min-w-0 mt-4 xl:order-first xl:mt-0">
       <slot />
+      <div v-if="$scopedSlots['prev-next']" class="px-4 sm:px-6">
+        <hr class="mt-10 mb-8 border-gray-100 dark:border-gray-800 dark:border-opacity-50" />
+        <slot name="prev-next" />
+      </div>
     </article>
 
     <slot name="toc" />
