+++ src/defaultTheme/components/templates/Docs.vue
@@ -1,10 +1,16 @@
 <template>
   <AppPage>
-    <PageContent :page="page" />
+    <PageContent :page="page">
+      <template #mobile-toc>
+        <PageMobileToc :title="page.toc.title" :toc="page.toc.links" />
+      </template>
+    </PageContent>
 
-    <hr class="mt-10 mb-8 border-gray-100 dark:border-gray-800 dark:border-opacity-50" />
+    <EditOnGithub :page="page" />
 
-    <PagePrevNext :prev="prev" :next="next" />
+    <template #prev-next>
+      <PagePrevNext :prev="prev" :next="next" />
+    </template>
 
     <template #toc>
       <PageToc v-if="!page.hideToc" :title="page.toc.title" :toc="page.toc.links" />
